+++ backend/src/controllers/auth/index.ts
@@ -1,1 +1,1 @@
-export * from './controller';
+export * from './auth-controller';
