+++ backend/src/controllers/book/auth-controller.ts
