+++ backend/src/controllers/book/index.ts
@@ -1,1 +1,1 @@
-export * from './controller';
+export * from './auth-controller';
