+++ backend/src/controllers/wishlist/index.ts
@@ -1,1 +1,1 @@
-export * from './controller';
+export * from './wishlist-controller';
