+++ backend/src/middleware/auth-middleware.ts
