+++ backend/src/services/auth/index.ts
@@ -1,1 +1,1 @@
-export * from './service';
+export * from './auth-service';
