+++ backend/src/services/books/index.ts
@@ -1,2 +1,2 @@
 export * from './types';
-export * from './service';
+export * from './books-service';
