+++ backend/src/services/wishlist/index.ts
@@ -1,2 +1,2 @@
 export * from './types';
-export * from './service';
+export * from './wishlist-service';
